fix: repository 정의 코드 추출
